Added the toolchain to the library, using the same toolchain to compile buildroot and kernel, solved the problem of buildroot and kernel mismatch, but the current toolchain can not compile kernel 4.19 and systemd cannot be configured, but the kernel can now be started

+++ devices/mkimage.sh
@@ -1,25 +1,61 @@
-#!/bin/sh
+#!/bin/bash
 set -e
+passwd='123'
 
 CURRENT_DIR=$(pwd)
 # 定义退出信号处理函数
 cleanup() {
-    if mount | grep -q "$CURRENT_DIR/p1"; then
-        sudo umount p1
-    fi
-    if mount | grep -q "$CURRENT_DIR/p2"; then
-        sudo umount p2
-    fi
-    if mount | grep -q "$CURRENT_DIR/rootfs"; then
-        sudo umount rootfs
-    fi
-    sudo losetup -d $loop_dev
+    # 搜索挂载的文件系统
+    cd $CURRENT_DIR
+    mount | grep "devices/p1" | awk '{print $3}' | while read -r mountpoint; do
+        echo "Unmounted $mountpoint"
+        sudo umount "$mountpoint"
+    done
+
+    mount | grep "devices/p2" | awk '{print $3}' | while read -r mountpoint; do
+        echo "Unmounted $mountpoint"
+        sudo umount "$mountpoint"
+    done
+    
+    mount | grep "rootfs" | awk '{print $3}' | while read -r mountpoint; do
+        echo "Unmounted $mountpoint"
+        sudo umount "$mountpoint"
+    done
+
+    loop_devices=$(losetup -a | grep "sd.img" | awk -F: '{print $1}')
+    for loop_device in $loop_devices; do
+        sudo losetup -d $loop_device
+    done
+
+    # if [ -d "$CURRENT_DIR/p1" ]; then
+    #     if mount | grep -q "$CURRENT_DIR/p1"; then
+    #         sudo umount p1
+    #         echo "umount p1"
+    #     fi
+    # fi
+    # if [ -d "$CURRENT_DIR/p2" ]; then
+    #     if mount | grep -q "$CURRENT_DIR/p2"; then
+    #         sudo umount p2
+    #         echo "umount p2"
+    #     fi
+    # fi
+    # if [ -d "$CURRENT_DIR/rootfs" ]; then
+    #     if mount | grep -q "$CURRENT_DIR/rootfs"; then
+    #         sudo umount rootfs
+    #         echo "umount rootfs"
+    #     fi
+    # fi
+    # if [ -n "$loop_dev" ]; then
+    #     sudo losetup -d $loop_dev
+    # fi
     sudo rm -rf p1 p2 rootfs
     sudo rm uImage
 }
 # 捕捉退出信号
 trap cleanup EXIT
 te_arch=$1
+
+echo $passwd | sudo -S ls
 
 kernel_path=./uImage
 # dtb_path=~/linux-4.19/kernel/arch/arm/boot/dts/te/te_vxpress.dtb
@@ -37,21 +73,26 @@ sec_img=sd.img
 
 ############### Create uImage
 cd ../u-boot/tools
-mkimage -n "virt_linux" -A arm -a 0x60008000 -e 0x60008000 \
- -d ../../kernel/arch/$te_arch/boot/Image ../../devices/uImage
+
+if [ ! -f sd.img ]; then
+    mkimage -n "virt_linux" -A arm -a 0x60008000 -e 0x60008000 \
+    -d ../../kernel/arch/$te_arch/boot/Image ../../devices/uImage
+fi
 
 cd $CURRENT_DIR
 ############### Create img
 # 这种只能arm架构不能使用arm64架构，还得继续看 https://blog.51cto.com/u_15127650/3467228
 loop_dev=$(losetup -f)
+echo "use $loop_dev"
 
-dd if=/dev/zero of=$sec_img bs=1024 count=524288
-
-#创建GPT分区，下面创建了两个分区，一个用来存放kernel和设备树，另一个存放根文件系统
-#sgdisk -n 0:0:+20M -c 0:kernel $sec_img
-#sgdisk -n 0:0:0 -c 0:rootfs $sec_img
-sgdisk -n 1:2048:43007 -c 1:"kernel" $sec_img   # kernel 分区
-sgdisk -n 2:43008:0 -c 2:"rootfs" $sec_img   # rootfs 分区
+if [ ! -f sd.img ]; then
+    dd if=/dev/zero of=$sec_img bs=1024 count=524288
+    #创建GPT分区，下面创建了两个分区，一个用来存放kernel和设备树，另一个存放根文件系统
+    #sgdisk -n 0:0:+20M -c 0:kernel $sec_img
+    #sgdisk -n 0:0:0 -c 0:rootfs $sec_img
+    sgdisk -n 1:2048:43007 -c 1:"kernel" $sec_img   # kernel 分区
+    sgdisk -n 2:43008:0 -c 2:"rootfs" $sec_img   # rootfs 分区
+fi
 
 sudo losetup $loop_dev $sec_img
 sudo partprobe $loop_dev
@@ -59,8 +100,8 @@ sudo partprobe $loop_dev
 #格式化
 #sudo mkfs.ext4 ${loop_dev}p1
 #sudo mkfs.ext4 ${loop_dev}p2
-sudo mkfs.ext4 -b 4096 -L kernel ${loop_dev}p1
-sudo mkfs.ext4 -b 4096 -L rootfs ${loop_dev}p2
+sudo mkfs.ext4 -b 4096 -F -L kernel ${loop_dev}p1
+sudo mkfs.ext4 -b 4096 -F -L rootfs ${loop_dev}p2
 
 sudo mkdir p1 p2 rootfs
 
@@ -75,7 +116,16 @@ if [ "$te_arch" = "arm" ]; then
     sudo cp -a $CURRENT_DIR/../kernel/arch/$te_arch/boot/zImage p1/
 fi
 sudo cp -a $dtb_path p1/
-sudo cp -raf rootfs/* ./p2
+if [ -s rootfs ]; then
+    sudo cp -raf rootfs/* ./p2
+else
+    echo "rootfs does not exist or is empty."
+fi
+
+cd p2/dev
+sudo mknod console c 5 1
+sudo mknod ttyAMA0 c 204 64
+
 #sudo cp -a $CURRENT_DIR/../rootfs/lib/* ./p2/lib
 
 # sudo umount p1 p2 rootfs
